requirements.txt
django-environ
requests
pyyaml
random-word